requirements.txt
python-telegram-bot==20.7
notion-client==2.2.1
flask[async]==3.0.2
python-dotenv==1.0.1
gunicorn==21.2.0
openai-whisper==20231117
schedule==1.2.1
python-dateutil==2.8.2
sentry-sdk[flask]==1.40.4
pytz==2023.3
timezonefinder==6.5.9
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
fastapi==0.110.0
uvicorn==0.27.1
